requirements.txt
numpy>=1.16.2
Pillow>=5.4.1
PyYAML>=5.1.2
